Revisions to throttle and drive_fsm based on comments by Arshan on Jan 3
