--- projects/pedal/inc/throttle_calibration.h
@@ -1,2 +1,0 @@
-#pragma once 
-
