--- projects/pedal/src/mech_brake.c
@@ -1,10 +1,0 @@
-#include "mech_brake.h" 
-
-#include <string.h> 
-#include <stdio.h> 
-#include <stdbool.h> 
-#include <limits.h>
-
-#include "event_queue.h"
-#include "log.h"
-//include "pedal_events.h"
