--- projects/pedal/src/throttle_calibration.c
@@ -1,10 +1,0 @@
-#include "mech_brake.h" // Change later 
-
-#include <string.h> 
-#include <stdio.h> 
-#include <stdbool.h> 
-#include <limits.h>
-
-#include "event_queue.h"
-#include "log.h"
-//include "pedal_events.h"
